[6.x] Split specifyParameter() to external trait (#31254)
* [6.x] Split specifyParameter() to external trait.

While building orchestra/canvas, I found out that while it would be nice
to use Illuminate\Console\Command the requirements to have Laravel (or
at least the Container) configure make it slightly harder to build PHP
CLI code based on symfony console with a little Laravel flavour.


* Remove unused import
